fix: clear lint findings in state, metrics and stacks
Three of these were real, not style:

- stacks/interfaces.go asserted `f.Interface().([]string)` after checking
  only `Elem().Kind() == String`. A named type such as `type Tags []string`
  passes that check but has a different dynamic type, so adding a stack
  field of one would have panicked at runtime. Now returns an error.

- state/user_cache.go dropped the error from `Close` on a *written* file.
  These files hold OAuth tokens, so a failed flush was being silently
  swallowed. The deferred close now surfaces its error when the function
  would otherwise have succeeded. The read path still ignores it, which is
  correct and now explicit.

- state/user_cache.go joined a caller-supplied name onto the cache path
  with no validation. Callers only ever pass literals, so nothing was
  exploitable, but the guard is three lines in code that writes tokens.

The rest:

- metrics.TimeFrame.Period returned int64 while every one of its four
  callers immediately converted to int32 for CloudWatch. Returning int32
  removes the conversions and the overflow warnings with them; the values
  are 60, 240 and 600.

- stacks/database.go: append(dst, src...) instead of a copy loop.

- stacks/cluster.go: net.DefaultResolver.LookupNS with a context instead of
  net.LookupNS, matching how the rest of the package calls out.

state, metrics and stacks now report zero findings.

+++ metrics/metrics.go
@@ -28,8 +28,11 @@ func (t *TimeFrame) Clone() *TimeFrame {
 	}
 }
 
-// Period returns the resolution in seconds to graph for the time frame
-func (t *TimeFrame) Period() int64 {
+// Period returns the resolution in seconds to graph for the time frame.
+//
+// int32 because that is what CloudWatch's MetricStat.Period takes; every
+// caller feeds it straight to aws.Int32.
+func (t *TimeFrame) Period() int32 {
 	switch {
 	case t.Duration.Hours() <= 24:
 		return 60
@@ -117,7 +120,7 @@ func (m *ServiceUtilizationMetrics) MetricDataQueries() []types.MetricDataQuery 
 						},
 					},
 				},
-				Period: aws.Int32(int32(m.Options.Timeframe.Period())),
+				Period: aws.Int32(m.Options.Timeframe.Period()),
 				Stat:   aws.String("Maximum"),
 			},
 		},
@@ -138,7 +141,7 @@ func (m *ServiceUtilizationMetrics) MetricDataQueries() []types.MetricDataQuery 
 						},
 					},
 				},
-				Period: aws.Int32(int32(m.Options.Timeframe.Period())),
+				Period: aws.Int32(m.Options.Timeframe.Period()),
 				Stat:   aws.String("Maximum"),
 			},
 		},
@@ -195,7 +198,7 @@ func (m *ResponseTimeMetrics) MetricDataQueries() []types.MetricDataQuery {
 						},
 					},
 				},
-				Period: aws.Int32(int32(m.Options.Timeframe.Period())),
+				Period: aws.Int32(m.Options.Timeframe.Period()),
 				Stat:   &m.Stat,
 			},
 		},
@@ -257,7 +260,7 @@ func (m *StatusCodeMetrics) MetricDataQueries() []types.MetricDataQuery {
 						},
 					},
 				},
-				Period: aws.Int32(int32(m.Options.Timeframe.Period())),
+				Period: aws.Int32(m.Options.Timeframe.Period()),
 				Stat:   aws.String("Sum"),
 			},
 		},
